+++ src/main.js
@@ -1,20 +1,17 @@
-import normalizeException from 'normalize-exception'
-
 import { mergeValues } from './merge.js'
 import { normalizeOptions } from './options.js'
 import { shouldSkipProp, shouldSetValue } from './skip.js'
 
 // Merge error properties.
 export default function setErrorProps(error, props, opts) {
-  const errorA = normalizeException(error)
-  const { lowPriority } = normalizeOptions(props, opts)
+  const { lowPriority } = normalizeOptions(error, props, opts)
 
   // eslint-disable-next-line fp/no-loops
   for (const propName of Reflect.ownKeys(props)) {
-    mergeProp({ error: errorA, props, propName, lowPriority })
+    mergeProp({ error, props, propName, lowPriority })
   }
 
-  return errorA
+  return error
 }
 
 // Just like `Object.assign()`: